+++ kernels/common/scene_triangle_mesh.cpp
@@ -171,7 +171,7 @@ namespace embree
 
   void TriangleMesh::postCommit() 
   {
-    scene->vertices[geomID] = (unsigned int*) vertices0.getPtr();
+    scene->vertices[geomID] = (float*) vertices0.getPtr();
 
     triangles.setModified(false);
     for (auto& buf : vertices)
